+++ projects/core/index.ts
@@ -56,6 +56,7 @@ export {
   IdMetadata,
   FindFirstOptionsBase,
 } from './src/remult3/remult3.js'
+export { standardSchema } from './src/standard-schema/index.js'
 export {
   EntityBase,
   ControllerBase,